services/ingest-git/requirements.txt
asyncpg>=0.27.0
tiktoken>=0.3.0
httpx>=0.24.0
python-dotenv>=1.0.0
pgvector>=0.4.0
gitpython>=3.1.40
